Web wasm node & channel history & updated contract & some refactor and features & frontend

+++ tonpayments/api/channel.go
@@ -16,7 +16,7 @@ import (
 )
 
 type Onchain struct {
-	CommittedSeqno uint32 `json:"committed_seqno"`
+	CommittedSeqno uint64 `json:"committed_seqno"`
 	WalletAddress  string `json:"wallet_address"`
 	Deposited      string `json:"deposited"`
 	Withdrawn      string `json:"withdrawn"`
